+++ vue/src/views/Header.vue
@@ -3,10 +3,10 @@
     <div class="header">
       <ul class="acount">
         <li class="personalinfo">
-          User Name：{{ account.username }}
+          User Name：{{ username }}
         </li>
         <li class="personalinfo">
-          Authority:{{ account.role }}
+          Authority:{{ role }}
         </li>
         <li>
           <span tag="button" class="btn" @click="logout()">
@@ -22,16 +22,14 @@
 export default {
   data () {
     return {
-      account: {
-        username: '',
-        role: ''
-      }
+      username: '',
+      role: ''
     }
   },
   async mounted () {
     const userId = this.$store.state.auth.userId
     // ロード時にactionsにdispatchする
-    await this.$store.dispatch('account/fetchAccount', { userId: userId })
+    await this.$store.dispatch('account/fetchAccount', userId)
     // storeから情報を取得するメソッド
     this.setAccount()
   },
@@ -39,15 +37,15 @@ export default {
     // dataのaccountにaccount.jsのstateの情報をsetする
     setAccount () {
       const stateAccount = this.$store.state.account
-      this.account.username = stateAccount.account.username
+      this.username = stateAccount.account.username
       const role = stateAccount.account.name
       // ROLE_USERだったら一般ユーザとヘッダーに表示される
       if (role === 'ROLE_USER') {
-        this.account.role = '一般ユーザ'
+        this.role = '一般ユーザ'
       } else if (role === 'ROLE_ADMIN') {
-        this.account.role = '管理者'
+        this.role = '管理者'
       } else {
-        this.account.role = 'オーナー'
+        this.role = 'オーナー'
       }
     },
     // ログアウトボタンが押された時のメソッド
